修复获取文件日志信息过程中日志文件被误删的bug

+++ FS_Server/src/main/java/com/bonree/brfs/disknode/data/write/record/RecordCollection.java
@@ -7,8 +7,11 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.List;
 
+import com.bonree.brfs.common.utils.CloseUtils;
 import com.bonree.brfs.common.utils.ProtoStuffUtils;
 import com.bonree.brfs.disknode.data.write.FileWriter;
 
@@ -21,15 +24,17 @@ import com.bonree.brfs.disknode.data.write.FileWriter;
 public class RecordCollection implements Closeable, Iterable<RecordElement> {
 	private File recordFile;
 	private FileWriter recordWriter;
-	private RecordCollectionManager manager;
+	private List<InputStream> openedStreams = new ArrayList<InputStream>();
+	
+	private boolean deleteOnClose;
 	
 	/**
 	 * 创建记录处理器
 	 * 
 	 * @param recordFilePath 记录文件路径
 	 */
-	RecordCollection(String recordFilePath, FileWriter fileWriter, RecordCollectionManager manager) {
-		this(new File(recordFilePath), fileWriter, manager);
+	RecordCollection(String recordFilePath, FileWriter fileWriter, boolean deleteOnClose) {
+		this(new File(recordFilePath), fileWriter, deleteOnClose);
 	}
 	
 	/**
@@ -37,10 +42,10 @@ public class RecordCollection implements Closeable, Iterable<RecordElement> {
 	 * 
 	 * @param recordFile 记录文件
 	 */
-	RecordCollection(File recordFile, FileWriter fileWriter, RecordCollectionManager manager) {
+	RecordCollection(File recordFile, FileWriter fileWriter, boolean deleteOnClose) {
 		this.recordFile = recordFile;
 		this.recordWriter = fileWriter;
-		this.manager = manager;
+		this.deleteOnClose = deleteOnClose;
 	}
 	
 	public File recordFile() {
@@ -71,6 +76,7 @@ public class RecordCollection implements Closeable, Iterable<RecordElement> {
 		InputStream input = null;
 		try {
 			input = new BufferedInputStream(new FileInputStream(recordFile));
+			openedStreams.add(input);
 		} catch (FileNotFoundException e) {
 			e.printStackTrace();
 		}
@@ -82,8 +88,14 @@ public class RecordCollection implements Closeable, Iterable<RecordElement> {
 	public void close() throws IOException {
 		//写入记录关闭的时候说明记录已经不需要了，可以将
 		//其关闭
-		recordFile.delete();
-		manager.releaseCollection(RecordFileBuilder.reverse(recordFile));
+		CloseUtils.closeQuietly(recordWriter);
+		for(InputStream input : openedStreams) {
+			CloseUtils.closeQuietly(input);
+		}
+		
+		if(deleteOnClose) {
+			recordFile.delete();
+		}
 	}
 	
 	private class RecordElementIterator implements Iterator<RecordElement> {
